Don't assume tools will be in <li> tags.
Some admin skins, such as Bootstrap-based ones, do not place object
tools in unordered lists. This change moves the wrapping <li> tags to
the templates, so that it may be overridden.

Also, a second, optional, argument was added to the 'object_tool'
templatetag to allow for templates to override the default argument for
the 'class' attribute on tool anchor tags.